arrange prototypes

+++ libft/ft_calloc.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_calloc(int n, int size)
+void	*ft_calloc(size_t n, size_t size)
 {
 	void	*save;
 	void	*p;
